new branch customers

+++ database/migrations/2021_03_01_111337_create_customers_table.php
@@ -17,7 +17,7 @@ class CreateCustomersTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('address');
-            $table->unsignedInteger('nip');
+            $table->unsignedBigInteger('nip');
             $table->timestamps();
         });
     }
